requirements.txt
streamlit
openai
pandas
faiss-cpu
numpy
langchain
pytest
logging